fix comment component and model
