+++ models/comments.js
@@ -1,15 +1,15 @@
 const db = require('../lib/dbConnect.js');
 
-function getAllComment(req, res, next) {
-  db.any(`SELECT *
-          FROM comment;
-          `)
-  .then((comments) => {
-    res.comments = comments;
-    next();
-  })
-  .catch(err => next(err));
-}
+// function getAllComment(req, res, next) {
+//   db.any(`SELECT *
+//           FROM comment;
+//           `)
+//   .then((comments) => {
+//     res.comments = comments;
+//     next();
+//   })
+//   .catch(err => next(err));
+// }
 
 function getComment(req, res, next) {
   db.one(`SELECT *
